requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.32.0,<1.0.0
google-adk>=1.0.0
google-cloud-bigquery>=3.35.0
google-api-python-client>=2.170.0
google-auth>=2.38.0
python-dateutil>=2.9.0
dateparser>=1.2.0
pydantic>=2.9.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
pytest>=8.3.0,<9.0.0
